# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImportGamesBody {
  username: string;
  /** Number of past months to fetch (1-12) */
  months?: number;
}

export interface ImportGamesResponse {
  imported: number;
  total: number;
  username: string;
}

export interface Game {
  id: number;
  username: string;
  pgn: string;
  whiteUsername: string;
  blackUsername: string;
  whiteRating: number;
  blackRating: number;
  result: string;
  timeControl: string;
  /** @nullable */
  opening?: string | null;
  /** @nullable */
  eco?: string | null;
  playedAt: string;
  /** @nullable */
  url?: string | null;
  analyzed: boolean;
  /** @nullable */
  analysisNotes?: string | null;
  createdAt: string;
}

export interface Move {
  moveNumber: number;
  san: string;
  color: string;
  /** @nullable */
  fen?: string | null;
  /** @nullable */
  comment?: string | null;
}

export interface GameReplay {
  id: number;
  pgn: string;
  moves: Move[];
  whiteUsername: string;
  blackUsername: string;
  whiteRating: number;
  blackRating: number;
  result: string;
  /** @nullable */
  opening?: string | null;
  /** @nullable */
  analysisNotes?: string | null;
}

export interface ListGamesResponse {
  games: Game[];
  total: number;
}

export interface AnalyzeGamesBody {
  username: string;
}

export interface Weakness {
  category: string;
  severity: string;
  description: string;
  frequency: number;
  examples: string[];
}

export interface AnalysisResult {
  username: string;
  gamesAnalyzed: number;
  weaknesses: Weakness[];
  summary: string;
}

export interface WeaknessRecord {
  id: number;
  username: string;
  category: string;
  severity: string;
  description: string;
  frequency: number;
  examples: string[];
  createdAt: string;
}

export interface WeaknessReport {
  username: string;
  weaknesses: WeaknessRecord[];
  /** @nullable */
  lastUpdated?: string | null;
}

export interface OpeningStat {
  opening: string;
  games: number;
  wins: number;
  losses: number;
  draws: number;
}

export interface TimeControlStat {
  timeControl: string;
  games: number;
  wins: number;
  losses: number;
}

export interface PerformanceSummary {
  username: string;
  totalGames: number;
  wins: number;
  losses: number;
  draws: number;
  winRate: number;
  avgRating: number;
  openingStats: OpeningStat[];
  resultsByTimeControl: TimeControlStat[];
}

export interface GenerateCoursesBody {
  username: string;
}

export interface Course {
  id: number;
  username: string;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  totalLessons: number;
  completedLessons: number;
  createdAt: string;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  content: string;
  orderIndex: number;
  completed: boolean;
  /** @nullable */
  examplePgn?: string | null;
}

export interface CourseWithLessons {
  id: number;
  username: string;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  totalLessons: number;
  completedLessons: number;
  createdAt: string;
  lessons: Lesson[];
}

export interface ListCoursesResponse {
  courses: Course[];
}

export interface UpdateProgressBody {
  lessonId: number;
  completed: boolean;
}

export type ListGamesParams = {
  /**
   * Filter by username
   */
  username?: string;
  /**
   * Limit results
   */
  limit?: number;
  /**
   * Offset for pagination
   */
  offset?: number;
};

export type GetWeaknessesParams = {
  username: string;
};

export type GetAnalysisSummaryParams = {
  username: string;
};

export type ListCoursesParams = {
  username: string;
};