# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  username: string;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  totalLessons: number;
  completedLessons: number;
  createdAt: string;
}