# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: number;
  username: string;
  pgn: string;
  whiteUsername: string;
  blackUsername: string;
  whiteRating: number;
  blackRating: number;
  result: string;
  timeControl: string;
  /** @nullable */
  opening?: string | null;
  /** @nullable */
  eco?: string | null;
  playedAt: string;
  /** @nullable */
  url?: string | null;
  analyzed: boolean;
  /** @nullable */
  analysisNotes?: string | null;
  createdAt: string;
}